4.21

+++ part4/controllers/blogs.js
@@ -10,14 +10,6 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs)
 })
 
-// blogsRouter.post('/', async (request, response) => {
-//   const blog = new Blog(request.body)
-
-//   const result = await blog.save()
-//   response.status(201).json(result)
-// })
-
-
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
@@ -25,9 +17,8 @@ blogsRouter.post('/', async (request, response) => {
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
-  
-  const user = await User.findById(decodedToken.id)
 
+  const user = await User.findById(decodedToken.id)
   if(!user){
     return response.status(400).json({ error: 'userId missing or not valid' })
   }
@@ -48,8 +39,27 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  if(!user){
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' });
+  }
+
+  if(blog.user.toString() === user.id.toString() ){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(403).json({ error: 'permission denied: only creator can delete' });
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
